backend/requirements-railway.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
boto3==1.35.0
botocore==1.35.99
build==1.3.0
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.3.0
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
deepdiff==8.6.1
distro==1.9.0
durationpy==0.10
emoji==2.15.0
exceptiongroup==1.3.0
fastapi==0.115.4
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==34.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.2
Markdown==3.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orderly-set==5.5.0
orjson==3.11.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pillow==12.0.0
pluggy==1.6.0
posthog==7.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.1
pydantic-settings==2.12.0
pydantic_core==2.27.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-iso639==2025.11.11
python-magic==0.4.27
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.15.3
sentence-transformers==2.5.1
shellingham==1.5.4
simpy==4.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.41.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.12.5
unstructured-client==0.25.9
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xgboost==2.0.3
yarl==1.22.0
zipp==3.23.0
